% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NCEI earthquake data}
\usage{
eq_clean_data(filename)
}
\arguments{
\item{filename}{A character string giving the path and name of the delimited
file to be read in e.g."data/earthquakes.tsv"}
}
\value{
Output will be a data.frame object.
}
\description{
Reads, tidies and returns earthquake data downloadedfrom the US National
Centers for Environmental Information:
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}.
Longitude and latitude fields are read in automatically as numeric data types
using the readr::read_delim function and do not need to be specified.
}
\examples{
\dontrun{
eq_clean_data("data/earthquakes.tsv")
}

}
