% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label() HTML label function}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{Data frame}
}
\value{
Output will be a vector of character strings
}
\description{
Builds eq_create_label() HTML label function for when generating the
Leaflet map
}
