% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans LOCATION_NAME column}
\usage{
eq_location_clean(dataframe)
}
\arguments{
\item{dataframe}{A character string giving the name of the input dataframe.}
}
\value{
Output will be a data.frame object.
}
\description{
Cleans LOCATION_NAME column and returns earthquake data from the US National Centers for
Environmental Information:
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}.
}
\examples{
\dontrun{
eq_location_clean(eq_clean_data("data/earthquakes.tsv"))
}

}
