% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{eq_map}
\alias{eq_map}
\title{eq_map() mapping function}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{Data frame}

\item{annot_col}{Attribute of data to be displayed as a popup on the map}
}
\value{
Output will be a Leaflet map
}
\description{
Builds the eq_map() mapping function for generating a Leaflet map
}
