% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{ggplot2 layer geom_timeline_label()}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetics mappings for call to ggplot2::layer}

\item{data}{Data to be displayed}

\item{stat}{Statistical transformation to be used}

\item{position}{Position adjustment}

\item{na.rm}{use na.rm = FALSE from function}

\item{show.legend}{Logical to include legends}

\item{inherit.aes}{Logical to overide default aesthetics}

\item{...}{list of other default parameters}
}
\value{
Output will be ggplot2 layer
}
\description{
Builds the geom_timeline_label() ggplot2 layer for the GeomTimeline geom.
}
