% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{ggproto class GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\value{
Output will be ggplot2 layer
}
\description{
Builds the ggproto class GeomTimeline for geom_timeline() funtion
}
\keyword{datasets}
