% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline()
}
\value{
Output will be ggplot2 layer
}
\description{
Function to build ggplot2 theme to display plot as specified in exercise
}
