% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{ggproto class GeomTimeline}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabel
}
\arguments{
\item{required_aes}{A character vector of aesthetics needed to render the geom}

\item{default_aes}{A list generated by aes() of default values for aesthetics}

\item{draw_key}{Renders a single legend key}

\item{draw_panel}{Method that assembles panels into a gtable object}
}
\value{
Output will be ggplot2 layer
}
\description{
Builds the ggproto class GeomTimeline for geom_timeline() funtion
}
\keyword{datasets}
