% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_proj.R
\name{theme_timeline}
\alias{theme_timeline}
\title{theme_timeline}
\usage{
theme_timeline()
}
\arguments{
\item{plot.background}{plot background colour}

\item{panel.background}{panel background colour}

\item{legend.key}{background underneath legend}

\item{axis.title.y}{y axis title}

\item{axis.line.x}{x axis line}

\item{axis.ticks.y}{ticks on the y axis}

\item{legend.position}{position of legend}
}
\value{
Output will be ggplot2 layer
}
\description{
Function to build ggplot2 theme to display plot as specified in exercise
}
